Move the notification service call out of the stress test demo
This was testing code that lasted for about 6 months without issue.
If we run ci while the notification service is deploying this will fail.
