+++ packages/stress/esbuild.config.mjs
@@ -5,6 +5,7 @@ build({
     entryPoints: {
         start: './src/start.ts',
         demo: './src/demo.ts',
+        notifications: './src/notifications.ts',
     },
     bundle: true,
     sourcemap: 'inline',
