+++ packages/stress/package.json
@@ -7,6 +7,7 @@
         "build": "yarn typecheck && yarn build-esbuild",
         "build-esbuild": "node esbuild.config.mjs",
         "demo": "node --enable-source-maps ./dist/demo.cjs",
+        "notifications": "node --enable-source-maps ./dist/notifications.cjs",
         "lint": "eslint . --ext .ts",
         "start": "node --enable-source-maps ./dist/start.cjs",
         "test:ci:with-entitlements": "vitest run --config ./vitest.config.entitlements.ts --silent",
